pyproject.toml
[tool.poetry]
name = "sitraved"
version = "0.1.0"
description = ""
authors = ["Cheiras"]

[tool.poetry.dependencies]
python = "^3.7"
Django = "^3.1.3"
django-environ = "^0.4.5"
psycopg2-binary = "^2.8.6"
djangorestframework = "^3.12.2"
pytest-django = "^4.1.0"
djangorestframework-simplejwt = "^4.6.0"
factory-boy = "^3.2.0"
django-cors-headers = "^3.7.0"
responses = "^0.13.2"
gunicorn = "^20.1.0"
flake8 = "^4.0.1"
[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
